Move Encodable* structs for Category and Keyword to the `views` module
