+++ src/views/mod.rs
@@ -1,12 +1,43 @@
 // TODO: Move all encodable types here
 // For now, just reexport
 
+use chrono::NaiveDateTime;
+
 pub use badge::EncodableBadge;
-pub use category::{EncodableCategory, EncodableCategoryWithSubcategories};
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct EncodableCategory {
+    pub id: String,
+    pub category: String,
+    pub slug: String,
+    pub description: String,
+    #[serde(with = "::util::rfc3339")] pub created_at: NaiveDateTime,
+    pub crates_cnt: i32,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct EncodableCategoryWithSubcategories {
+    pub id: String,
+    pub category: String,
+    pub slug: String,
+    pub description: String,
+    #[serde(with = "::util::rfc3339")] pub created_at: NaiveDateTime,
+    pub crates_cnt: i32,
+    pub subcategories: Vec<EncodableCategory>,
+}
+
 pub use crate_owner_invitation::{EncodableCrateOwnerInvitation, InvitationResponse};
 pub use dependency::EncodableDependency;
 pub use download::EncodableVersionDownload;
-pub use keyword::EncodableKeyword;
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct EncodableKeyword {
+    pub id: String,
+    pub keyword: String,
+    #[serde(with = "::util::rfc3339")] pub created_at: NaiveDateTime,
+    pub crates_cnt: i32,
+}
+
 pub use krate::EncodableCrate;
 pub use owner::{EncodableOwner, EncodableTeam};
 pub use token::EncodableApiTokenWithToken;
